Move payment method view model
